+++ service/src/main/java/org/ehrbase/dao/access/util/FolderUtils.java
@@ -2,13 +2,46 @@ package org.ehrbase.dao.access.util;
 
 import com.nedap.archie.rm.datastructures.ItemStructure;
 import com.nedap.archie.rm.directory.Folder;
+import org.ehrbase.api.exception.ObjectNotFoundException;
 import org.ehrbase.dao.access.interfaces.I_FolderAccess;
 import org.ehrbase.serialisation.RawJson;
+import org.jooq.JSONB;
 import org.postgresql.util.PGobject;
+
+import java.util.Optional;
 
 public class FolderUtils {
 
-    private FolderUtils() {
+    private FolderUtils() {}
+
+    public static I_FolderAccess getPath(I_FolderAccess folderAccess, int currentIndex, String[] path) {
+
+        // End of recursion since we have reached the end of the path
+        if (currentIndex >= path.length) {
+            return folderAccess;
+        }
+
+        String searchFolderName = path[currentIndex];
+
+        // Throw not found on empty list of folders
+        if (folderAccess.getSubfoldersList() == null || folderAccess.getSubfoldersList().isEmpty()) {
+            throw new ObjectNotFoundException("DIRECTORY", "Folder at path '" + searchFolderName + "' could not be found.");
+        }
+
+        Optional<I_FolderAccess> foundFolder = folderAccess
+                .getSubfoldersList()
+                .values()
+                .stream()
+                .filter(subFolderAccess -> searchFolderName.equals(subFolderAccess.getFolderName()))
+                .findFirst();
+
+        // Throw not found if folder could not be found
+        if (foundFolder.isEmpty()) {
+            throw new ObjectNotFoundException("DIRECTORY", "Folder at path '" + searchFolderName + "' could not be found." );
+        }
+
+        // Call recursive search for next path element
+        return getPath(foundFolder.get(), currentIndex + 1, path);
     }
 
     /**
@@ -33,6 +66,14 @@ public class FolderUtils {
             target.getItems().clear();
             target.getItems().addAll(update.getItems());
         }
+    }
+
+    public static ItemStructure parseFromJSONB(JSONB dbObject) {
+        if (dbObject == null) {
+            return null;
+        }
+        String value = dbObject.data();
+        return new RawJson().unmarshal(value, ItemStructure.class);
     }
 
     public static ItemStructure parseFromPGobject(PGobject databaseObject) {