+++ validation/src/main/java/org/ehrbase/validation/terminology/Pathables.java
@@ -5,11 +5,13 @@ import com.nedap.archie.rm.archetyped.Pathable;
 import org.apache.commons.lang3.StringUtils;
 import org.ehrbase.terminology.openehr.TerminologyInterface;
 import org.ehrbase.terminology.openehr.implementation.AttributeCodesetMapping;
+import org.ehrbase.validation.terminology.validator.ItemField;
 
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
 import java.lang.reflect.Field;
+import java.util.List;
 
 public class Pathables {
 
@@ -18,7 +20,7 @@ public class Pathables {
     private AttributeCodesetMapping codesetMapping;
     private String language;
 
-    public Pathables(TerminologyInterface terminologyInterface, AttributeCodesetMapping codesetMapping, ItemValidator itemValidator, String language) {
+    Pathables(TerminologyInterface terminologyInterface, AttributeCodesetMapping codesetMapping, ItemValidator itemValidator, String language) {
         this.terminologyInterface = terminologyInterface;
         this.itemValidator = itemValidator;
         this.codesetMapping = codesetMapping;
@@ -28,31 +30,43 @@ public class Pathables {
     public void traverse(Pathable pathable, String... excludes) throws IllegalArgumentException, InternalError {
 
         for (Field field: pathable.getClass().getDeclaredFields()){
-            try {
-                if (field.getType() == field.getType().asSubclass(Pathable.class)) {
+            if (!field.getType().equals(List.class)) {
+                try {
+                    if (field.getType() == field.getType().asSubclass(Pathable.class)) {
 
-                    if (isFieldExcluded(excludes, field.getName()))
-                        continue;
+                        if (isFieldExcluded(excludes, field.getName()))
+                            continue;
 
-                    Object object = objectForField(pathable, field);
+                        RMObject object = new ItemField<RMObject>(pathable).objectForField(field);
 
-                    if (object != null && object instanceof Pathable) {
-                        new Pathables(terminologyInterface, codesetMapping, itemValidator, language).traverse((Pathable) object, excludes);
+                        if (object instanceof Pathable) {
+                            new Pathables(terminologyInterface, codesetMapping, itemValidator, language).traverse((Pathable) object, excludes);
+                        } else if (object != null)
+                            throw new IllegalArgumentException("Internal: couldn't handle object retrieved using getter");
+                    }
+                } catch (ClassCastException e) {
+                    //check if object is handled for validation
+                    if (itemValidator.isValidatedRmObjectType(field.getType())) {
+                        RMObject object =new ItemField<RMObject>(pathable).objectForField(field);
+                        itemValidator.validate(terminologyInterface, codesetMapping, field.getName(), object, language);
+                    }
+                }
+            }//continue
+            else { //iterate the array
+                List iterable = new ItemField<List>(pathable).objectForField(field);
+
+                for (Object item: iterable){
+                    if (item instanceof RMObject){
+                        itemValidator.validate(terminologyInterface, codesetMapping, field.getName(), (RMObject)item, language);
+                    }
+                    else if (item instanceof Pathable){
+                        traverse((Pathable)item, excludes);
                     }
                     else
-                        if (object != null)
-                            throw new IllegalArgumentException("Internal: couldn't handle object retrieved using getter");
+                        throw new IllegalStateException("Could not handle item in list:"+item);
                 }
             }
-            catch (ClassCastException e){
-                //check if object is handled for validation
-                if (itemValidator.isValidatedRmObjectType(field.getType())){
-                    RMObject object = objectForField(pathable, field);
-                    itemValidator.validate(terminologyInterface, codesetMapping, field.getName(), object, language);
-                }
-            } //continue
         }
-
     }
 
     private boolean isFieldExcluded(String[] excludes, String fieldName){
@@ -63,26 +77,4 @@ public class Pathables {
 
         return false;
     }
-
-    private RMObject objectForField(Pathable pathable, Field field) throws IllegalArgumentException, InternalError {
-        String getterName = "get"+ StringUtils.capitalize(field.getName());
-        MethodHandle methodHandle;
-        try {
-            methodHandle = MethodHandles.lookup().findVirtual(pathable.getClass(), getterName, MethodType.methodType(field.getType()));
-        }
-        catch (NoSuchMethodException | IllegalAccessException e){
-            throw new InternalError("Internal error:"+e.getMessage());
-        }
-        try {
-            Object object = methodHandle.invoke(pathable);
-            if (object != null && !(object instanceof RMObject))
-                throw new IllegalArgumentException("Internal: object is not of class RMObject:" + object.toString());
-
-            return (RMObject) object;
-        }
-        catch (Throwable throwable){
-            throw new InternalError("Internal:"+throwable.getMessage());
-        }
-    }
-
 }